+++ neulbomgil-backend/src/main/java/com/gooroomees/neulbomgil_backend/domain/chat/controller/ChatController.java
@@ -17,12 +17,12 @@ public class ChatController {
     private final ChatService chatService;
 
     @PostMapping("/start")
-    public ChatRoomResponseDto startChatRoom(@RequestParam Integer userId) {
+    public ChatRoomResponseDto startChatRoom(@RequestParam Long userId) {
         return chatService.startChatRoom(userId);
     }
 
     @GetMapping("/{roomId}/message")
-    public List<ChatResponseDto> getMessagesByRoomId(@PathVariable Integer roomId){
+    public List<ChatResponseDto> getMessagesByRoomId(@PathVariable Long roomId){
 return chatService.getMessageByRoomId(roomId);
     }
 
@@ -32,12 +32,12 @@ return chatService.getMessageByRoomId(roomId);
     }
 
     @PostMapping("/{roomId}/message")
-    public ChatResponseDto sendMessage(@PathVariable Integer roomId,@RequestBody ChatRequestDto requestDto) {
+    public ChatResponseDto sendMessage(@PathVariable Long roomId,@RequestBody ChatRequestDto requestDto) {
         return chatService.saveMessage(roomId, requestDto);
     }
 
     @PutMapping("/{roomId}/read")
-    public void readMessages(@PathVariable Integer roomId, @RequestParam Integer userId) {
+    public void readMessages(@PathVariable Long roomId, @RequestParam Long userId) {
         chatService.readMessages(roomId,userId);
     }
