+++ neulbomgil-backend/src/main/java/com/gooroomees/neulbomgil_backend/domain/chat/dto/ChatRoomResponseDto.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @Builder
 public class ChatRoomResponseDto {
-    private Integer roomId;
+    private Long roomId;
 
     private Long userId;
